requirements.txt
azure-functions==1.21.3
azure-data-tables==12.5.0
python-telegram-bot==21.9
feedparser==6.0.11
requests==2.32.3
python-dotenv==1.0.1
